separate V1 and V2 interpreters and adjust tests accordingly
